Move the extension to a separate file
