+++ DuckDuckGo/SubscriptionDebugViewController.swift
@@ -347,28 +347,3 @@ import NetworkProtection
         }
     }
 }
-
-extension UINavigationController {
-
-    func popToRootViewController(animated: Bool = true, completion: @escaping ()->()) {
-        CATransaction.begin()
-        CATransaction.setCompletionBlock(completion)
-        self.popToRootViewController(animated: animated)
-        CATransaction.commit()
-    }
-
-    func popViewControllerWithHandler(animated: Bool = true, completion: @escaping ()->()) {
-        CATransaction.begin()
-        CATransaction.setCompletionBlock(completion)
-        self.popViewController(animated: animated)
-        CATransaction.commit()
-    }
-
-    func pushViewController(viewController: UIViewController, animated: Bool = true,  completion: @escaping ()->()) {
-        CATransaction.begin()
-        CATransaction.setCompletionBlock(completion)
-        self.pushViewController(viewController, animated: animated)
-        CATransaction.commit()
-    }
-
-}
